# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sonart Insight API",
    "description": "\nERP ve reklam raporlarini yonetim dashboard'una ceviren analiz servisi.\n\n**Mimarinin iki tasiyici fikri:**\n\n1. **Dataset Pack** — boru hattinin tamami (okuma, kodlama onarimi, veri\n   kalitesi, zaman serisi, risk tespiti, AI) rapor tipinden bagimsizdir. Bir\n   rapora ozgu her sey tek bir yapilandirma nesnesinde toplanir. Yeni rapor\n   eklemek = yeni pack yazmak.\n\n2. **Sayiyi Polars hesaplar, AI yorumlar.** Modele ham veri hic verilmez;\n   yalnizca hesaplanmis tablolar gider. Uretilen her aksiyon sema geregi kanit\n   tasimak zorundadir ve her kanit hesaplanmis degerle karsilastirilarak\n   dogrulanir (`grounding_ratio`).\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasets": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Yüklenmiş veri setleri",
        "operationId": "list_datasets_api_v1_datasets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetSummary"
                  },
                  "type": "array",
                  "title": "Response List Datasets Api V1 Datasets Get"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "CSV yükle ve işle",
        "operationId": "upload_dataset_api_v1_datasets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_datasets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Veri seti özeti",
        "operationId": "get_dataset_api_v1_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSummary"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "datasets"
        ],
        "summary": "Veri setini sil",
        "operationId": "delete_dataset_api_v1_datasets__dataset_id__delete",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/quality": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Veri kalitesi raporu",
        "description": "Tespit edilen her sorun; ağırlığı, kaç satırı etkilediği ve **ne yapıldığı** (onarildi / turetildi / silindi / karantina / isaretlendi) ile birlikte.",
        "operationId": "get_quality_api_v1_datasets__dataset_id__quality_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityReportOut"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/overview": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Dashboard özeti",
        "operationId": "get_overview_api_v1_datasets__dataset_id__overview_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "top",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Dönülecek risk sayısı",
              "default": 8,
              "title": "Top"
            },
            "description": "Dönülecek risk sayısı"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/periods": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Dönemsel trend ve değişim verisi",
        "operationId": "get_periods_api_v1_datasets__dataset_id__periods_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeriodsResponse"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/entities": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Kayıt bazında özet ve zaman serisi",
        "operationId": "get_entities_api_v1_datasets__dataset_id__entities_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "include_series",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "(kayıt, dönem) uzun tablosunu da döndür",
              "default": true,
              "title": "Include Series"
            },
            "description": "(kayıt, dönem) uzun tablosunu da döndür"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitiesResponse"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/risks": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Risk sicili",
        "operationId": "get_risks_api_v1_datasets__dataset_id__risks_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ağırlığa göre filtre, ör. 'kritik'",
              "title": "Severity"
            },
            "description": "Ağırlığa göre filtre, ör. 'kritik'"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Yalnızca bu dönemde açılan riskler",
              "title": "Period"
            },
            "description": "Yalnızca bu dönemde açılan riskler"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RisksResponse"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_dataset_api_v1_datasets_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "İşlenecek CSV dosyası"
          },
          "pack": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack",
            "description": "Rapor tipi (ör. 'sonart-erp'). Boş bırakılırsa başlıklardan tespit edilir."
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_api_v1_datasets_post"
      },
      "DatasetSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "pack_key": {
            "type": "string",
            "title": "Pack Key"
          },
          "pack_title": {
            "type": "string",
            "title": "Pack Title"
          },
          "raw_row_count": {
            "type": "integer",
            "title": "Raw Row Count"
          },
          "clean_row_count": {
            "type": "integer",
            "title": "Clean Row Count"
          },
          "quarantined_row_count": {
            "type": "integer",
            "title": "Quarantined Row Count"
          },
          "health_score": {
            "type": "integer",
            "title": "Health Score"
          },
          "encoding_detected": {
            "type": "string",
            "title": "Encoding Detected"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "pack_key",
          "pack_title",
          "raw_row_count",
          "clean_row_count",
          "quarantined_row_count",
          "health_score",
          "encoding_detected",
          "created_at"
        ],
        "title": "DatasetSummary"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "schema_mismatch"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Tum hatalarin ortak govdesi.\n\nIstemci `code` alanina gore dallanir; mesaj metni degisse bile kirilmaz."
      },
...
      "OverviewResponse": {
        "properties": {
          "dataset": {
            "$ref": "#/components/schemas/DatasetSummary"
          },
          "periods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Periods"
          },
          "entity_count": {
            "type": "integer",
            "title": "Entity Count"
          },
          "headline_metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricOut"
            },
            "type": "array",
            "title": "Headline Metrics"
          },
          "period_rows": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Period Rows",
            "description": "Dönem bazında portföy özeti -- trend grafiklerinin kaynağı."
          },
          "risk_counts_by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Risk Counts By Severity"
          },
          "top_risks": {
            "items": {
              "$ref": "#/components/schemas/RiskOut"
            },
            "type": "array",
            "title": "Top Risks"
          }
        },
        "type": "object",
        "required": [
          "dataset",
          "periods",
          "entity_count",
          "headline_metrics",
          "period_rows",
          "risk_counts_by_severity",
          "top_risks"
        ],
        "title": "OverviewResponse"
      },
...
      "QualityReportOut": {
        "properties": {
          "raw_row_count": {
            "type": "integer",
            "title": "Raw Row Count"
          },
          "clean_row_count": {
            "type": "integer",
            "title": "Clean Row Count"
          },
          "quarantined_row_count": {
            "type": "integer",
            "title": "Quarantined Row Count"
          },
          "encoding_detected": {
            "type": "string",
            "title": "Encoding Detected"
          },
          "encoding_repaired": {
            "type": "boolean",
            "title": "Encoding Repaired"
          },
          "health_score": {
            "type": "integer",
            "title": "Health Score"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/QualityIssueOut"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "raw_row_count",
          "clean_row_count",
          "quarantined_row_count",
          "encoding_detected",
          "encoding_repaired",
          "health_score",
          "issues"
        ],
        "title": "QualityReportOut"
      },
...
      "UploadResponse": {
        "properties": {
          "dataset": {
            "$ref": "#/components/schemas/DatasetSummary"
          },
          "quality": {
            "$ref": "#/components/schemas/QualityReportOut"
          },
          "periods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Periods"
          },
          "entity_count": {
            "type": "integer",
            "title": "Entity Count"
          },
          "risk_count": {
            "type": "integer",
            "title": "Risk Count"
          }
        },
        "type": "object",
        "required": [
          "dataset",
          "quality",
          "periods",
          "entity_count",
          "risk_count"
        ],
        "title": "UploadResponse"
      }
    }
  }
}